requirements.txt
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
Flask==2.2.2
flask-restx==1.0.3
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
marshmallow==3.19.0
packaging==21.3
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytz==2022.6
SQLAlchemy==1.4.44
Werkzeug==2.2.2
